Address comments, move beta contact adding to contactsStoreUtil
